Upgrade dependencies: * com.maxmind.geoip2 * org.apache.commons:commons-compress * com.google.guava:guava * org.apache.httpcomponents.client5:httpclient5 * netty-*
+ enable some tests only when MAXMIND_LICENSE_KEY env var is set

+++ common-geoip/src/main/java/be/dnsbelgium/mercator/geoip/DisabledGeoIPService.java
@@ -18,12 +18,12 @@ public class DisabledGeoIPService implements GeoIPService {
   }
 
   @Override
-  public Optional<Pair<Integer, String>> lookupASN(InetAddress ip) {
+  public Optional<Pair<Long, String>> lookupASN(InetAddress ip) {
     return Optional.empty();
   }
 
   @Override
-  public Optional<Pair<Integer, String>> lookupASN(String ip) {
+  public Optional<Pair<Long, String>> lookupASN(String ip) {
     return Optional.empty();
   }
 }
